chore: move filterPetSitter api to petSitterController, delete filterController
